Homework10 moved CorpDataParser to parse_utils; added complete test for both classes in package; mocked collect_info in MarketsInsiderRawData to awoid calls to web; found web-dependency in CorpDataParser._get_rate - mocked it and added fixture for it
